c316: extract BenchmarksAxes tab cluster (CrossSceneTransfer + RateDistortion + MutualInfo + EndmemberBaseline + SpatialCoherence + SuperTopics) to own file (#441 P0 2.2)
Benchmarks.tsx 1620 LOC -> 935 LOC (-685 LOC). Benchmarks chunk 35.83 KB -> 18.56 KB (-17.3 KB).
New chunk: BenchmarksAxes-CqpOxiel.js 17.83 KB.
typecheck clean, 44/44 tests pass.
